@@ -1,8 +1,8 @@
 package info.romanelli.udacity.android.popularmovies.util;
 
-import android.support.annotation.NonNull;
-import android.support.v4.view.ViewCompat;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.core.view.ViewCompat;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -29,7 +29,7 @@ public class MovieInfoAdapter extends RecyclerView.Adapter<MovieInfoAdapter.View
     }
 
     public void setData(final List<MovieInfo> listMovieInfo) {
-        this.listMovieInfo = (listMovieInfo != null) ? listMovieInfo : new ArrayList<MovieInfo>();
+        this.listMovieInfo = (listMovieInfo != null) ? listMovieInfo : new ArrayList<>();
         notifyDataSetChanged();
     }
 
