@@ -1,7 +1,7 @@
 package info.romanelli.udacity.android.popularmovies.database;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.ViewModel;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.ViewModel;
 import android.util.Log;
 
 import info.romanelli.udacity.android.popularmovies.util.AppDatabase;
